Introduced PageSelector class for page selection during insertRecord
Create PageSelector for each file opened with rbfm instance.
Whenever we have to insertRecord, select which page to insert the
record to using this.
This class maintains pageOccupancy info as heap and gets the page
with available space to insert the record. Else it creates a new
page, and returns that page.
All metadata are stored as hidden pages that are managed by the
PageSelector class
Added method in pfm to set hidden pages used by rbfm layer