Changed how we store Attribute type on the disk
The attribute type of the index key is stored as an int It's represented
by enum in the code, but on the disk I'm storing it as int.

Previously we were storing it as Attribute struct directly which had
string, enum and integer. And after deserialising and converting that to
the struct, we were getting munmap_chunk invalid free error. This was
very weird, since the cause wasn't obvious. But now while writing the
commit message I'm realising we never accounted for Attribute.name
string variable. Didn't properly serialize/deserialize that. And didn't
even account for it's length. It can be variable length.

So changed that to store just integer which can be converted to enum
upon deserializing