Fix elements being reset during insertion
While inserting an entry into vector of RidAndKey, if the insertIdx is
somewhere in the middle of the vector, the elements after the insertIdx
including the insertIdx had their keys reset. This was weird behavior
So, instead of inserting using vector.insert function, make a copy
of the vector and copy the elements into the vector.

The same thing happened with vector.erase for vector of RidAndKey.
So changed the logic in delete function also